refactor(web): decompose session-sidebar into hook + subcomponents
